gcp/requirements.txt
torch
transformers==4.20.0
grpcio
grpcio-tools
asyncio
asyncpg
google-cloud-aiplatform
cloud-sql-python-connector
langchain
pgvector
gradio
bs4
datasets>=2.2.0
pyarrow==8.0.0
jsonlines
numpy
requests>=2.26.0
scikit_learn>=1.0.2
scipy
sentence_transformers>=2.2.0
tqdm
rich